ama documentation
